iprt/asm*,++: Contain the intrin.h mess in iprt/win/msvc_intrin.h.  Hack it for bs3kit and no-crt to avoid dragging in malloc.h from UCRT with VC++ 14.1.  bugref:8489
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@83776 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/VBox/vmm/ssm.h
@@ -237,7 +237,7 @@ typedef struct SSMFIELD
 /** Emit a SSMFIELD array entry.
  * @internal  */
 #define SSMFIELD_ENTRY_INT(Name, off, cb, enmTransformer, uFirstVer) \
-    { (PFNSSMFIELDGETPUT)(uintptr_t)(enmTransformer), (off), (cb), (uFirstVer), Name }
+    { (PFNSSMFIELDGETPUT)(uintptr_t)(enmTransformer), (uint32_t)(off), (uint32_t)(cb), (uFirstVer), Name }
 /** Emit a SSMFIELD array entry.
  * @internal  */
 #define SSMFIELD_ENTRY_TF_INT(Type, Field, enmTransformer, uFirstVer) \
